fix: (c++) 22193 Multiply (40점)

+++ 22100~22199/22193.cpp
@@ -18,7 +18,7 @@
 using namespace std;
 
 int main() {
-  int n, m, a, b;
+  long long n, m, a, b;
   cin >> n >> m >> a >> b;
   cout << a * b << '\n';
   return 0;
